missing 'i' in bash script

+++ data/genfiles.sh
@@ -2,14 +2,16 @@
 
 chown -R 1010:1010 *
 chmod -R 775 *
-f [ -e "./etc-nginx.tar.gz" ]
+if [ -e "./etc-nginx.tar.gz" ]
 then
+  echo "Deleting etc-nginx.tar.gz"
   rm etc-nginx.tar.gz
 fi
 cd etc-nginx
 tar -c * | gzip -f > ../etc-nginx.tar.gz
-f [ -e "$http.tar.gz" ]
+if [ -e "$http.tar.gz" ]
 then
+  echo "Deleting http.tar.gz"
   rm http.tar.gz
 fi
 cd http
